+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- **ext:** *(experimental)* An argument list with type assertions
+
+
 ## 0.8.0 - 2017-08-06
 
 ### Added